requirements.txt
asgiref==3.8.1
black==24.2.0
boto3==1.37.29
click==8.1.8
crispy-bootstrap5==0.7
Django==5.2
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-storages==1.14.6
djangorestframework==3.14.0
flake8==7.0.0
gunicorn==21.2.0
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.3.7
pluggy==1.5.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
pytest==7.4.4
pytest-django==4.7.0
python-dotenv==1.0.0
pytz==2025.2
sqlparse==0.5.3
whitenoise==6.6.0
